requirements.txt
markdown2
bleach
requests
google-generativeai
Django>=5.2.6
tensorflow==2.18.0
keras==3.8.0
pillow
matplotlib
django-tinymce
python-dotenv
